+++ src/lib/formats/fs_cbmdos.h
@@ -13,10 +13,7 @@
 
 #pragma once
 
-
 #include "fsmgr.h"
-#include <optional>
-#include <string_view>
 
 namespace fs {
