+++ src/lib/formats/fs_coco_os9.h
@@ -14,8 +14,6 @@
 #pragma once
 
 #include "fsmgr.h"
-#include <optional>
-#include <string_view>
 
 namespace fs {
