+++ src/lib/formats/fs_hplif.h
@@ -13,10 +13,7 @@
 
 #pragma once
 
-
 #include "fsmgr.h"
-#include <optional>
-#include <string_view>
 
 namespace fs {
 
@@ -38,7 +35,6 @@ namespace fs {
 		virtual std::vector<meta_description> volume_meta_description() const override;
 		virtual std::vector<meta_description> file_meta_description() const override;
 	};
-
 
 	extern const hplif_image HPLIF;
