+++ src/lib/formats/fs_unformatted.cpp
@@ -4,6 +4,8 @@
 // Creation of unformatted floppy images
 
 #include "fs_unformatted.h"
+#include "flopimg.h"
+#include "fsblk.h"
 
 namespace fs {
